WIP: Working with results #2

+++ app/src/main/java/com/gtohelper/domain/models/SportResult.kt
@@ -4,7 +4,7 @@ data class SportResult(
     val id: Int? = null,
     val sportName: String,
     val competitionId: Int,
-    val value: Long,
+    val value: Int,
     val competitorNumber: Int,
 )
